back to int

+++ Stimulator/Parameters.h
@@ -16,17 +16,17 @@ NS_ASSUME_NONNULL_BEGIN
 @interface Parameters : NSObject {
     @public int _blocks_wide;
     @public int _blocks_high;
-    @public int _block_height;
-    @public int _block_width;
+    @public float _block_height;
+    @public float _block_width;
     @public float _stoplight_time;
-    @public int _street_width;
+    @public float _street_width;
     @public PolicyFunc _policy;
 
     @public int _min_time;
     @public int _max_time;
 };
 
-- (instancetype)initWithBlocksWide:(int)blocksWide BlocksHigh: (int)blocksHigh blockHeight: (int)blockHeight blockWidth: (int)blockWidth stoplightTime: (int)stoplightTime streetWidth: (int)streetWidth policy: (PolicyFunc)policy;
+- (instancetype)initWithBlocksWide:(int)blocksWide BlocksHigh:(int)blocksHigh blockHeight:(float)blockHeight blockWidth:(float)blockWidth stoplightTime:(float)stoplightTime streetWidth:(float)streetWidth policy: (PolicyFunc)policy;
 
 @end
